--- crates/gen-guest-c/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    println!("cargo:rerun-if-changed=build.rs");
-    // this build script is currently only here so OUT_DIR is set for testing.
-}
