+++ crates/gen-guest-rust/Cargo.toml
@@ -17,4 +17,5 @@ clap = { workspace = true, optional = true }
 
 [dev-dependencies]
 wit-bindgen-guest-rust = { path = '../guest-rust' }
-test-helpers = { path = '../test-helpers', features = ['guest-rust'] }
+guest-rust-test-macro = { path = 'test-macro' }
+test-helpers = { path = '../test-helpers', default-features = false }
