+++ crates/gen-guest-rust/test-macro/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "guest-rust-test-macro"
+version = "0.0.0"
+edition.workspace = true
+publish = false
+
+[lib]
+proc-macro = true
+test = false
+doctest = false
+
+[dependencies]
+wit-parser = { workspace = true }
+proc-macro2 = "1.0.27"
+quote = "1.0.9"
+heck = { workspace = true }
