+++ crates/host-wasmtime-rust-macro/src/lib.rs
@@ -62,6 +62,7 @@ mod kw {
     syn::custom_keyword!(src);
     syn::custom_keyword!(paths);
     syn::custom_keyword!(custom_error);
+    syn::custom_keyword!(tracing);
 }
 
 impl Parse for Opts {
@@ -79,6 +80,7 @@ impl Parse for Opts {
             for field in fields.into_pairs() {
                 match field.into_value() {
                     ConfigField::Interfaces(v) => interfaces = v,
+                    ConfigField::Tracing(v) => opts.tracing = v,
                     ConfigField::CustomError(v) => opts.custom_error = v,
                 }
             }
@@ -114,6 +116,7 @@ impl Parse for Opts {
 enum ConfigField {
     Interfaces(Vec<Interface>),
     CustomError(bool),
+    Tracing(bool),
 }
 
 impl Parse for ConfigField {
@@ -151,6 +154,10 @@ impl Parse for ConfigField {
             Ok(ConfigField::CustomError(
                 input.parse::<syn::LitBool>()?.value,
             ))
+        } else if l.peek(kw::tracing) {
+            input.parse::<kw::tracing>()?;
+            input.parse::<Token![:]>()?;
+            Ok(ConfigField::Tracing(input.parse::<syn::LitBool>()?.value))
         } else {
             Err(l.error())
         }
