+++ crates/test-helpers/macros/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "test-helpers-macros"
+authors = ["Alex Crichton <alex@alexcrichton.com>"]
+version.workspace = true
+edition.workspace = true
+publish = false
+
+[lib]
+proc-macro = true
+doctest = false
+test = false
+
+[dependencies]
+backtrace = "0.3"
+heck = { workspace = true }
+ignore = "0.4"
+proc-macro2 = "1.0.27"
+quote = "1.0.9"
+wit-bindgen-core = { workspace = true }
+wit-parser = { workspace = true }
+filetime = "0.2"
+
+[build-dependencies]
+wit-bindgen-gen-guest-c = { workspace = true }
+wit-bindgen-gen-guest-teavm-java = { workspace = true }
+wit-bindgen-core = { workspace = true }
+wit-component = { workspace = true }
+
+[features]
+guest-rust = []
+guest-c = []
+guest-teavm-java = []
